+++ CHANGELOG.md
@@ -5,8 +5,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
-### Changes
+### Additions
+- add layout to watched files (#17)
 - expose `sitedown.watch` method
+
+### Changes
 - use `path.resolve` instead of `cwp`
 
 ## [3.0.1](https://github.com/ungoldman/sitedown/releases/v3.0.1) - 2016-09-17