+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## [3.1.0](https://github.com/ungoldman/sitedown/releases/v3.1.0) - 2016-11-11
+
+[view diff](https://github.com/ungoldman/sitedown/compare/v3.0.1...v3.1.0)
 
 ### Additions
 - add layout to watched files (#17)